refactor: state패키지 내부 패키지명 전체를 소문자로 구성
