+++ src/main/java/blackjack/model/state/running/Hit.java
@@ -1,9 +1,9 @@
-package blackjack.model.state.Running;
+package blackjack.model.state.running;
 
 import blackjack.model.card.Card;
 import blackjack.model.card.Cards;
-import blackjack.model.state.Finished.Bust;
-import blackjack.model.state.Finished.Stay;
+import blackjack.model.state.finished.Bust;
+import blackjack.model.state.finished.Stay;
 import blackjack.model.state.State;
 
 public class Hit extends Running {
